Add borked cars

+++ src/lane.hpp
@@ -3,11 +3,11 @@
 #include <vector>
 
 struct Lane {
-    float left, width;
+    int left, width;
 };
 
 namespace Game {
     inline bool no_lanes_bitch = false;
-    inline float break_lane_width;
+    inline int break_lane_width;
     inline std::vector<Lane> front_lanes, back_lanes;
 } // namespace Game
